+++ src/com/zh/base/model/bean/RoleMenu.java
@@ -1,23 +1,23 @@
 package com.zh.base.model.bean;
 
 public class RoleMenu {
-	private Integer roleId;
+	private Long roleId;
 	
-	private Integer menuId;
+	private Long menuId;
 
-	public Integer getRoleId() {
+	public Long getRoleId() {
 		return roleId;
 	}
 
-	public void setRoleId(Integer roleId) {
+	public void setRoleId(Long roleId) {
 		this.roleId = roleId;
 	}
 
-	public Integer getMenuId() {
+	public Long getMenuId() {
 		return menuId;
 	}
 
-	public void setMenuId(Integer menuId) {
+	public void setMenuId(Long menuId) {
 		this.menuId = menuId;
 	}
